added user authentication to content. It is attached to specific userId now

+++ backend/src/controllers/routes.ts
@@ -2,14 +2,20 @@ import { RequestHandler } from "express";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from "../models/note";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async function (
   req,
   res,
   next
 ) {
+  const authenticatedUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authenticatedUserId);
+
+    const notes = await NoteModel.find({
+      userId: authenticatedUserId,
+    }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -23,6 +29,9 @@ export const getNote: RequestHandler = async function (
 ) {
   try {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
+
+    assertIsDefined(authenticatedUserId);
 
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
@@ -32,6 +41,10 @@ export const getNote: RequestHandler = async function (
 
     if (!note) {
       throw createHttpError(404, "note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "not authorized. No access");
     }
 
     res.status(200).json(note);
@@ -54,12 +67,16 @@ export const createNotes: RequestHandler<
   try {
     const title = req.body.title;
     const text = req.body.title;
+    const authenticatedUserId = req.session.userId;
+
+    assertIsDefined(authenticatedUserId);
 
     if (!title) {
       throw createHttpError(400, "Note must have a title");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -89,6 +106,10 @@ export const updateNotes: RequestHandler<
     const newTitle = req.body.title;
     const newText = req.body.text;
 
+    const authenticatedUserId = req.session.userId;
+
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "invalid note id");
     }
@@ -100,6 +121,10 @@ export const updateNotes: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "not authorized. No access");
     }
 
     note.title = newTitle;
@@ -119,7 +144,12 @@ export const deleteNote: RequestHandler = async function (
   next
 ) {
   const noteId = req.params.noteId;
+
+  const authenticatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
@@ -128,6 +158,10 @@ export const deleteNote: RequestHandler = async function (
 
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "not authorized. No access");
     }
 
     await note.deleteOne({ _id: noteId });
